Naval Expertise Official Federation of Teachers
The  U.S. Navy was born and bred to protect and serve, whether over seas or here on homeland soil. This is, and always was our purpose, and we will always back our personel and our country's families, and friends. This is why we have been called upon by our commander and our chief of the country, Mr. Obama and our concealed chief of staff. In combination, these two minds have been driven mad by the science it takes to right the wrongs, and errors we have found within our medical-providers' befallen services. By law you cannot, and will not supress the value of a life in our U.S. country, or abroad. This is our purpose for service over seas, and is always our purpose to defend, and end any victimization from malpractice or poaching of our U.S. veterans, and their beloved ones. This is non-VA affiliated, any and all complaints of our interference of your services are not to be directed to the VA because the VA is extremely busy cleaning up the backlisted mass of veteran patients already compiled by past ongoing corruption within the medical-provider community here in homeland. We are an independent company of non-corporation basis. We treat all our beloved with full care, empathy, love and respect their privacy at the cost of our death. We take privacy seriously, if privacy is infringed upon in and by any means; you will be subject to the judicial court for a hearing, and will face fining without limitation by judicials' judge.